+++ apps/app/src/test/scala/com/daml/network/integration/tests/DFDirectoryIntegrationTest.scala
@@ -31,8 +31,6 @@ class DFDirectoryIntegrationTest extends CNNodeIntegrationTest with WalletTestUt
 
   import DFDirectoryIntegrationTest.*
 
-  private val directoryDarPath =
-    "daml/directory-service/.daml/dist/directory-service-0.1.0.dar"
   private val testEntryName = "mycoolentry.unverified.cns"
   private val testEntryUrl = "https://cns-dir-url.com"
   private val testEntryDescription = "Sample CNS Directory Entry Description"
@@ -46,10 +44,6 @@ class DFDirectoryIntegrationTest extends CNNodeIntegrationTest with WalletTestUt
         CNNodeConfigTransforms.onlySv1
       )
       .withTrafficTopupsEnabled
-      .withAdditionalSetup(implicit env => {
-        aliceValidatorBackend.participantClient.upload_dar_unless_exists(directoryDarPath)
-        bobValidatorBackend.participantClient.upload_dar_unless_exists(directoryDarPath)
-      })
 
   "Directory service" should {