+++ apps/app/src/test/scala/com/daml/network/integration/tests/WalletFrontendTimeBasedIntegrationTest.scala
@@ -73,7 +73,7 @@ class WalletFrontendTimeBasedIntegrationTest
 
     "show user details after login" in { implicit env =>
       val aliceDamlUser = aliceWalletClient.config.ledgerApiUser
-      val aliceParty = setupForTestWithDirectory(aliceWalletClient, aliceValidatorBackend)
+      val aliceParty = onboardWalletUser(aliceWalletClient, aliceValidatorBackend)
       val entryName = perTestCaseName("alice")
 
       createDirectoryEntry(