+++ apps/app/src/test/scala/com/daml/network/integration/tests/WalletTimeBasedIntegrationTest.scala
@@ -24,8 +24,6 @@ class WalletTimeBasedIntegrationTest
     with SplitwellTestUtil {
 
   private val splitwellDarPath = "daml/splitwell/.daml/dist/splitwell-0.1.0.dar"
-  private val directoryDarPath =
-    "daml/directory-service/.daml/dist/directory-service-0.1.0.dar"
   private val testEntryName = "mycoolentry.unverified.cns"
   private val testEntryUrl = "https://cns-dir-url.com"
   private val testEntryDescription = "Sample CNS Directory Entry Description"
@@ -38,8 +36,6 @@ class WalletTimeBasedIntegrationTest
       .withAdditionalSetup(implicit env => {
         aliceValidatorBackend.participantClient.upload_dar_unless_exists(splitwellDarPath)
         bobValidatorBackend.participantClient.upload_dar_unless_exists(splitwellDarPath)
-        aliceValidatorBackend.participantClient.upload_dar_unless_exists(directoryDarPath)
-        bobValidatorBackend.participantClient.upload_dar_unless_exists(directoryDarPath)
       })
 
   "A wallet" should {