+++ apps/app/src/test/scala/com/daml/network/integration/tests/WalletTransactionHistoryFrontendTimeBasedIntegrationTest.scala
@@ -32,7 +32,7 @@ class WalletTransactionHistoryFrontendTimeBasedIntegrationTest
 
     "show all subscription payments" in { implicit env =>
       val aliceDamlUser = aliceWalletClient.config.ledgerApiUser
-      val aliceUserParty = setupForTestWithDirectory(aliceWalletClient, aliceValidatorBackend)
+      val aliceUserParty = onboardWalletUser(aliceWalletClient, aliceValidatorBackend)
       val aliceEntryName = perTestCaseName("alice")
 
       val directoryExpectedCns = createDirectoryEntryForDirectoryItself