+++ apps/app/src/test/scala/com/daml/network/integration/tests/runbook/SelfHostedPreflightIntegrationTest.scala
@@ -74,9 +74,6 @@ class SelfHostedPreflightIntegrationTest
       runScript(validatorPath / "validator.sc")(env.environment)
       runScript(validatorPath / "tap-transfer-demo.sc")(env.environment)
 
-      v("validatorApp").participantClient.dars
-        .upload("./daml/directory-service/.daml/dist/directory-service-0.1.0.dar")
-
       val walletUiPort = 3000
       val directoryUiPort = 3004