+++ apps/validator/src/main/scala/com/daml/network/validator/ValidatorApp.scala
@@ -15,6 +15,7 @@ import com.daml.network.admin.http.{HttpAdminHandler, HttpErrorHandler}
 import com.daml.network.auth.*
 import com.daml.network.codegen.java.cc.v1test as ccV1Test
 import com.daml.network.codegen.java.cn.appmanager.store as appManagerCodegen
+import com.daml.network.codegen.java.cn.directory as directoryCodegen
 import com.daml.network.codegen.java.cn.wallet.install as installCodegen
 import com.daml.network.config.{NetworkAppClientConfig, SharedCNNodeAppParameters}
 import com.daml.network.environment.*
@@ -155,13 +156,19 @@ class ValidatorApp(
       participantAdminConnection: ParticipantAdminConnection
   ): Future[Unit] = {
     logger.info(s"Attempting to setup wallet...")
-    val darFiles = new UploadablePackage {
-      lazy val packageId: String =
-        installCodegen.WalletAppInstall.TEMPLATE_ID.getPackageId
+    val darFiles = Seq(
+      new UploadablePackage {
+        lazy val packageId: String =
+          installCodegen.WalletAppInstall.TEMPLATE_ID.getPackageId
 
-      // See `Compile / resourceGenerators` in build.sbt
-      lazy val resourcePath: String = "dar/wallet-0.1.0.dar"
-    } +: Seq(new UploadablePackage {
+        // See `Compile / resourceGenerators` in build.sbt
+        lazy val resourcePath: String = "dar/wallet-0.1.0.dar"
+      },
+      new UploadablePackage {
+        lazy val packageId: String = directoryCodegen.DirectoryInstall.TEMPLATE_ID.getPackageId
+        lazy val resourcePath: String = "dar/directory-service-0.1.0.dar"
+      },
+    ) ++ Seq(new UploadablePackage {
       // should be the same as package dependency in wallet app
       lazy val packageId: String =
         ccV1Test.coin.CoinRulesV1Test.COMPANION.TEMPLATE_ID.getPackageId